\name{Delta}
\alias{Delta}
\docType{data}
\title{
List of Relatedness Coefficients
}
\description{
A list of relatedness coefficients obtained with the example of \cite{RelCoef}.
}
\usage{data("Delta")}
\format{
  The format is:
List of 15
 $ Delta1 : num [1:4, 1:4] 0.1872 0 0 0 0.0199 ...
 $ Delta2 : num [1:4, 1:4] 0.00 0.00 0.00 0.00 3.55e-05 ...
 $ Delta3 : num [1:4, 1:4] 0 0 0 0 0.0472 ...
 $ Delta4 : num [1:4, 1:4] 0 0 0 0 0.0322 ...
 $ Delta5 : num [1:4, 1:4] 0 0 0 0 0.0871 ...
 $ Delta6 : num [1:4, 1:4] 0 0 0 0 0.0395 ...
 $ Delta7 : num [1:4, 1:4] 0 0 0 0 0.0429 ...
 $ Delta8 : num [1:4, 1:4] 0 0 0 0 0.0386 ...
 $ Delta9 : num [1:4, 1:4] 0.8128 0 0 0 0.0202 ...
 $ Delta10: num [1:4, 1:4] 0 0 0 0 0.000731 ...
 $ Delta11: num [1:4, 1:4] 0 0 0 0 0.028 ...
 $ Delta12: num [1:4, 1:4] 0 0 0 0 0.0849 ...
 $ Delta13: num [1:4, 1:4] 0 0 0 0 0.0174 ...
 $ Delta14: num [1:4, 1:4] 0 0 0 0 0.0437 ...
 $ Delta15: num [1:4, 1:4] 0 0 0 0 0.498 ...
}
\examples{
data(Delta)
print(Delta$Delta7)
}
\keyword{datasets}
