\name{Frequencies}
\alias{Frequencies}
\docType{data}
\title{
Allele Frequencies
}
\description{
The allele frequencies matrix for the example with 5000 markers and one population.
}
\usage{data("Frequencies")}
\format{
  The format is:
 num [1:5000, 1:2] 0.268 0.786 0.804 0.238 0.235 ...
}
\examples{
data(Frequencies)
head(Frequencies)
}
\keyword{datasets}
