\name{Genotype}
\alias{Genotype}
\docType{data}
\title{
Genotype Matrix
}
\description{
The Parental Line Genom matrix for the example with 10 parental lines genotyped with 5000 markers.
}
\usage{data("Genotype")}
\format{
  The format is:
 num [1:5000, 1:10] 0 1 0 0 0 1 1 0 1 1 ...
}
\examples{
data(Genotype)
head(Genotype)
}
\keyword{datasets}
